+++ src/analysis.h
@@ -30,5 +30,22 @@
 
 void log_event_groups(nfer_specification *, dictionary *);
 
+/*** Compute specification cycles ***/
+typedef struct _rule_digraph_vertex {
+    nfer_rule     *rule;
+    unsigned int  index;    /* unique index of the vertex */
+    unsigned int  lowlink;  /* smallest index of vertex on stack known to be reachable */
+    bool          on_stack; /* is the vertex on the stack */
+    bool          visited;  /* track if the vertex has been visited yet */
+} rule_digraph_vertex;
+
+typedef struct _rule_digraph_edge {
+    struct _rule_digraph_vertex    *from;
+    struct _rule_digraph_vertex    *to;
+} rule_digraph_edge;
+
+bool compute_sccs(nfer_rule *, rule_digraph_vertex *, unsigned int, rule_digraph_edge *, unsigned int);
+bool setup_rule_order(nfer_specification *);
+bool exclusive_cycle(nfer_specification *);
 
 #endif /* INC_ANALYSIS_H_ */